abandon cmake. Do all jobs in bash.

+++ build.sh
@@ -17,22 +17,56 @@ if ! hash jq; then
     sudo apt install -yqq jq
 fi
 
+TARGET_TAG="collabccci/$OS_NAME:$OS_VER"
+
 # Check if we needed build
 # Use skopeo project
-sudo docker pull alexeiled/skopeo
-base="$(sudo docker run --rm alexeiled/skopeo skopeo inspect "docker://docker.io/$OS_NAME:$OS_VER" | jq -r '.Layers[-1]')"
-expect="$(sudo docker run --rm alexeiled/skopeo skopeo inspect "docker://docker.io/collabccci/$OS_NAME:$OS_VER" | jq -r '.Layers[-2]')"
-if [ "$base" = "$expect" ]; then
-    echo "Image $OS_NAME:$OS_VER is latest. Do nothing."
-    exit 0
+if [ "$TRAVIS_EVENT_TYPE" = "cron" ]; then
+    sudo docker pull alexeiled/skopeo
+    base="$(sudo docker run --rm alexeiled/skopeo skopeo inspect "docker://docker.io/$OS_NAME:$OS_VER" | jq -r '.Layers[-1]')"
+    expect="$(sudo docker run --rm alexeiled/skopeo skopeo inspect "docker://docker.io/collabccci/$OS_NAME:$OS_VER" | jq -r '.Layers[-2]')"
+    if [ "$base" = "$expect" ]; then
+        echo "Image $OS_NAME:$OS_VER is latest. Do nothing."
+        exit 0
+    fi
 fi
+
 
 
 echo "$DOCKER_PASSWORD" | sudo docker login --username "collabcc" --password-stdin
 
+
 rm -rf build
 mkdir build
 cd build
-cmake ..
-cmake --build .
+
+case "$OS_NAME" in
+    "ubuntu"|"debian"):
+        cp ../Dockerfile.apt ./Dockerfile
+        ;;
+    "centos"|"fedora"):
+        cp ../Dockerfile.yum ./Dockerfile
+        ;;
+    *):
+        echo "Unknown OS_NAME: $OS_NAME"
+        exit 1
+        ;;
+esac
+
+# Do substitution
+set +x
+env | while IFS=$'\n' read -r envvar; do
+    IFS='=' read -r name value <<<"$envvar"
+    name=${name//\\/\\\\}
+    value=${value//\\/\\\\}
+    #echo "[$name][$value]"
+    sed -i "s/@${name//\//\\/}@/${value//\//\\/}/g" Dockerfile
+done
+set -x
+
+echo "Substitution done"
+cat Dockerfile
+
+sudo docker build --rm --force-rm --pull --tag "${TARGET_TAG}" .
+sudo docker push "${TARGET_TAG}"
